Replace IsURL by IsWebAddress
The purpose of IsURL was checking if a string was a web address
(http://www.adevinta.com) or not.

It was not working correctly because for example localhost:8080 was
being taken as a  valid absolute URL, and it was not its purpose.

We are adding are replacing the IsURL method by IsWebAddress but, in
order to not break existing integrations, the method IsURL will be kept
just encapusulating the new one.